% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie3.R
\name{evaluate.ranking.direct}
\alias{evaluate.ranking.direct}
\title{Title}
\usage{
evaluate.ranking.direct(value, trues, perf.measures = c("acc", "rec", "prec",
  "fpr", "spec", "phi", "f"))
}
\arguments{
\item{trues}{}
}
\description{
Title
}

