% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.R
\name{plot_curves}
\alias{plot_curves}
\title{Make a plot of the ROC and PR curves.}
\usage{
plot_curves(eval_object)
}
\value{
The evaluation AU plots.
}
\description{
Make a plot of the ROC and PR curves.
}
\seealso{
\code{\link{run_genie3}}
}

