% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genie3.R
\name{retrieve_genie3_output}
\alias{retrieve_genie3_output}
\title{Retrieve GENIE3 output}
\usage{
retrieve_genie3_output(handle)
}
\arguments{
\item{handle}{}
}
\description{
Necessary when retrieving the output in combination with the PRISM package.
}
\seealso{
\code{\link{run_genie3}}
}

