% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genie3.R
\name{evaluate.multiple.rankings}
\alias{evaluate.multiple.rankings}
\title{Title Evaluate and compare multiple rankings}
\usage{
evaluate.multiple.rankings(rankings, true.matrix, perf.measures = c("acc",
  "rec", "prec", "fpr", "spec", "phi", "f"))
}
\arguments{
\item{rankings}{a list of rankings as preduced by \code{\link{get.ranking}}. See \code{\link{evaluate.ranking}} for more information.}

\item{true.matrix}{a matrix with 0's and 1's, representing the golden standard. The rownames and colnames must me the same as the names used in the regulator and target columns in \code{ranking}.}

\item{perf.measures}{the ROCR performance measures (See \code{\link[ROCR]{performance}}). Must at least contain \code{"fpr"}, \code{"rec"}, \code{"spec"}, and \code{"prec"}.}
}
\value{
a list containing 2 items, the ranked evaluation and the area under the curve scores
}
\description{
Title Evaluate and compare multiple rankings
}
\seealso{
\code{\link{genie3}}
}

