% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie3.R
\name{evaluate.ranking.direct}
\alias{evaluate.ranking.direct}
\title{Evaluate a ranking}
\usage{
evaluate.ranking.direct(is.true, num.positives, num.negatives,
  num.extend.steps = 10000)
}
\arguments{
\item{is.true}{A vector with 0's and 1's, representing the golden standard.}

\item{num.positives}{The total number of positives}

\item{num.negatives}{The total number of genatives}

\item{num.extend.steps}{The number of steps with which to fill the ranking as if random, if only a part of the ranking is given}
}
\value{
a list containing two items, the ranked evaluation and the area under the curve scores
}
\description{
Evaluate a ranking
}

