% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genie3.R
\name{evaluate.ranking}
\alias{evaluate.ranking}
\title{Title Evaluate a network ranking}
\usage{
evaluate.ranking(ranking, true.matrix, perf.measures = c("acc", "rec", "prec",
  "fpr", "spec", "phi", "f"))
}
\arguments{
\item{ranking}{the data frame as returned by \code{\link{get.ranking}}. This data frame must contain at least 2 columns, called \code{regulator} and \code{target}.}

\item{true.matrix}{a matrix with 0's and 1's, representing the golden standard. The rownames and colnames must me the same as the names used in the regulator and target columns in \code{ranking}.}

\item{perf.measures}{the ROCR performance measures (See \code{\link[ROCR]{performance}}). Must at least contain \code{"fpr"}, \code{"rec"}, \code{"spec"}, and \code{"prec"}.}
}
\value{
a list containing 2 items, the ranked evaluation and the area under the curve scores
}
\description{
Title Evaluate a network ranking
}
\seealso{
\code{\link{genie3}}
}

