% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genie3.R
\name{get.ranking}
\alias{get.ranking}
\title{Make an ordered data frame of regulatory links}
\usage{
get.ranking(weights, max.links = 1e+05)
}
\arguments{
\item{weights}{A weighted adjacency matrix as returned by \code{\link{genie3}}}

\item{max.links}{The maximum number of links to output.}
}
\value{
A data frame of links.
}
\description{
Make an ordered data frame of regulatory links
}
\seealso{
\code{\link{genie3}}
}

